requirements.txt
Package                Version
---------------------- ----------
APScheduler            3.6.3
beautifulsoup4         4.9.3
certifi                2020.6.20
cffi                   1.14.3
chardet                3.0.4
cryptography           3.1.1
cycler                 0.10.0
dateparser             1.0.0
decorator              4.4.2
geopy                  2.0.0
GoogleNews             1.5.0
humanize               3.1.0
idna                   2.10
kiwisolver             1.2.0
matplotlib             3.3.1
mysql-connector-python 8.0.22
numpy                  1.19.2
pandas                 1.1.3
Pillow                 8.0.0
pip                    20.2.4
protobuf               3.13.0
pycparser              2.20
pyparsing              2.4.7
python-dateutil        2.8.1
python-telegram-bot    13.0
pytz                   2020.1
regex                  2020.10.28
requests               2.24.0
scipy                  1.5.2
setuptools             50.3.2
six                    1.15.0
soupsieve              2.0.1
tornado                6.0.4
tzlocal                2.1
urllib3                1.25.10
wheel                  0.35.1
